Rag_Project/requirements.txt
openai
python-dotenv
faiss-cpu
langchain
pypdf
scikit-learn
